+++ t/1144-isf-public-tested-by-metadata-audit.t
@@ -255,6 +255,7 @@ my $expected_tested_by = [
     't/1351-isf-activation-param-package-constants.t',
     't/1352-isf-library-use-package-constants.t',
     't/1357-isf-transaction-port-package-constant-widths.t',
+    't/1368-isf-transaction-port-transaction-param-widths.t',
     't/1358-isf-data-op-package-constant-widths.t',
     't/1359-isf-wait-package-constant-counts.t',
     't/1360-isf-repeat-package-constant-counts.t',